% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zlib.R
\name{zlib}
\alias{zlib}
\alias{.onLoad}
\title{.onLoad function for the package}
\usage{
.onLoad(libname, pkgname)
}
\description{
This function is automatically called when the package is loaded using
\code{library()} or \code{require()}. It initializes the package environment,
including defining a variety of constants related to the zlib compression
library.
}
\details{
Specifically, the function assigns a new environment named "zlib" containing
constants such as \code{DEFLATED}, \code{DEF_BUF_SIZE}, \code{MAX_WBITS},
and various flush and compression strategies like \code{Z_FINISH},
\code{Z_BEST_COMPRESSION}, etc.
}
\seealso{
\code{\link[=publicEval]{publicEval()}} for the method used to set up the public environment.

\code{\link[=zlib_constants]{zlib_constants()}} for the method used to set up the constants in the environment.
}
\keyword{internal}
